% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfilarJSON.R
\name{perfilarJSON}
\alias{perfilarJSON}
\title{Perfila potenciales clientes}
\usage{
perfilarJSON(inputFile)
}
\arguments{
\item{inputFile}{Ubicación del input JSON file.}
}
\value{
JSON data con los resultados del perfilado.
}
\description{
Esta función toma descriptores del potencial afiliado en formato JSON y
devuelve una predicción de baja temprana (indicada como el campo PREDICCION = 1).
}
